+++ apiphany-httpclient5/src/main/java/org/apiphany/http/ApacheHC5Library.java
@@ -19,9 +19,15 @@ public class ApacheHC5Library {
 	public static final String CLIENT_NAME = "http-client5";
 
 	/**
+	 * The Apache HttpClient 5 circular redirect exception class name.
+	 */
+	protected static final String CIRCULAR_REDIRECT_EXCEPTION_CLASS_NAME =
+			"org.apache.hc.client5.http.CircularRedirectException";
+
+	/**
 	 * The Apache HttpClient 5 CloseableHttpClient class name.
 	 */
-	private static final String APACHE_HC5_CLOSEABLE_HTTP_CLIENT_CLASS_NAME =
+	private static final String CLOSEABLE_HTTP_CLIENT_CLASS_NAME =
 			"org.apache.hc.client5.http.impl.classic.CloseableHttpClient";
 
 	/**
@@ -30,7 +36,7 @@ public class ApacheHC5Library {
 	 * @return {@code true} if the Apache HttpClient 5 library is present, {@code false} otherwise
 	 */
 	public static boolean isPresent() {
-		return Reflection.isClassPresent(APACHE_HC5_CLOSEABLE_HTTP_CLIENT_CLASS_NAME);
+		return Reflection.isClassPresent(CLOSEABLE_HTTP_CLIENT_CLASS_NAME);
 	}
 
 	/**